Update djangoproject/templates/core/issue.html
Hi. Just as a test I've wrapped all the blocks with divs with sensible id-attributes. Now, nothing will happen until said ids are defined in the CSS-file too but it's a start. I do not currently have the means to run this on my PC so I'm just adding this as an example.