Reduce settings dependence on env_settings
The current strategy to configure the project's
env is to define a custom module out of the
project to hold env specific settings. Then the
actual settings imports it and use each param.

The usual way of doing this is to consider
settings.py as a production settings and then
**override** it's parameters with *env specific*
and *sensitive* values.

The **override** happens in the end of settings.