Add test for home viewed by user without profile.
I started this test because the home view has a check
to redirects an user without profile to complete his
registration.

However, during the process of writing this test, I
understood that this behavior is now resolved by a
middleware.

In a near future, this test should be moved to a 
middleware focused test.